Split DataLayerCard components into container and presentatio
